Add utils, cast types.
This is kind of a navel-gazing session for me, but I think it leaves the
code a little cleaner.  What's more, I think that I'm discovering
something (utils) which may be my route towards abstracting Draftable as
a general thing to manage different render targets.

Utils -> Render Engine?  Render Target?