Refactor.  Operations into separate file.
We've got a little bit of a refactor in order to decouple the Operation
engine from its pieces.  I can see a future in combining the Operations
file with the utils file, and just calling that "Engine".  Or if the
files can't combine, we'll at least have a good opportunity for a plugin
style engine bounded context.

Also in this commit, we split DraftableResult the type, and the class.
WTB a better name than Result, though.