Draftable section handling.
In order to expose control over a Draftable's sections (also known as a
Block Map in Draft lingo), we now have the 'sections' operation.

In essence, this lets the user run a single function over every section
(block) of a Draftable, producing a new Draftable with the results of
the function.

For example, if you wanted to change the text of each section to say
"Yo":

```javascript
Draftable
  .of("Two\nParagraphs")
  .sections(_ => Draftable.of("Yo"))
  .getText()

// "Yo\nYo"
```

As the Draftable DSL expands, this will hopefully create more enjoyable
and powerful transformation possibilities.

One caveat here is that we're using the EditorState.push feature, which
demands a pretty EditorState specific / Draft specific change "type".
Here we're using "change-block-data" as the type, so that's probably
going to result in some unintended consequences.

Finally, there's a clear opportunity here to change some terminology
(look for this as some tidiness refactors soon):

* Utils -> This is actually a Draft.js engine.  If we want to include
  mobilekit at some point as a render target, this is what we'd drop in
  a replacement for.  This should be called Engine, and given an
  interface.

* Operations / Operation, these need better names.  What's a good way to
  split these up?  An Operation is an object which results in a token,
  but the Operations object is the runtime that executes the token when
  it is finally invoked.  Operation -> Step?  Operations -> Brain?